// apps/api.k3h4.io/prisma/generated/schemas/objects/UsdaCommodityAttributeWhereInput.schema.ts
import * as z from 'zod';
import type { Prisma } from '@prisma/client';
import { StringFilterObjectSchema as StringFilterObjectSchema } from './StringFilter.schema';
import { StringNullableFilterObjectSchema as StringNullableFilterObjectSchema } from './StringNullableFilter.schema';
import { JsonNullableFilterObjectSchema as JsonNullableFilterObjectSchema } from './JsonNullableFilter.schema';
import { DateTimeFilterObjectSchema as DateTimeFilterObjectSchema } from './DateTimeFilter.schema'

const usdacommodityattributewhereinputSchema = z.object({
  AND: z.union([z.lazy(() => UsdaCommodityAttributeWhereInputObjectSchema), z.lazy(() => UsdaCommodityAttributeWhereInputObjectSchema).array()]).optional(),
  OR: z.lazy(() => UsdaCommodityAttributeWhereInputObjectSchema).array().optional(),
  NOT: z.union([z.lazy(() => UsdaCommodityAttributeWhereInputObjectSchema), z.lazy(() => UsdaCommodityAttributeWhereInputObjectSchema).array()]).optional(),
  id: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  dataset: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  code: z.union([z.lazy(() => StringFilterObjectSchema), z.string()]).optional(),
  name: z.union([z.lazy(() => StringNullableFilterObjectSchema), z.string()]).optional().nullable(),
  wikidataId: z.union([z.lazy(() => StringNullableFilterObjectSchema), z.string()]).optional().nullable(),
  enrichment: z.lazy(() => JsonNullableFilterObjectSchema).optional(),
  extra: z.lazy(() => JsonNullableFilterObjectSchema).optional(),
  createdAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional(),
  updatedAt: z.union([z.lazy(() => DateTimeFilterObjectSchema), z.coerce.date()]).optional()
}).strict();
export const UsdaCommodityAttributeWhereInputObjectSchema: z.ZodType<Prisma.UsdaCommodityAttributeWhereInput> = usdacommodityattributewhereinputSchema as unknown as z.ZodType<Prisma.UsdaCommodityAttributeWhereInput>;
export const UsdaCommodityAttributeWhereInputObjectZodSchema = usdacommodityattributewhereinputSchema;